# Optimize `take_bits` / `take_boolean`: up to -25%