--- a/arrow-select/src/take.rs
+++ b/arrow-select/src/take.rs
@@ -429,7 +429,7 @@ fn take_bits<I: ArrowPrimitiveType>(
         Some(nulls) => {
             let mut output_buffer = MutableBuffer::new_null(len);
             let output_slice: &mut [u8] = output_buffer.as_slice_mut();
-            output_buffer.nulls.valid_indices().for_each(|idx| {
+            nulls.valid_indices().for_each(|idx| {
                 if values.value(indices.value(idx).as_usize()) {
                     bit_util::set_bit(output_slice, idx);
                 }