# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #https://github.com/apache/arrow-rs/issues/6621

# Rationale for this change

Further optimization

```
take bool 1024          time:   [864.69 ns 865.71 ns 866.55 ns]
                        change: [-25.978% -25.819% -25.660%] (p = 0.00 < 0.05)
                        Performance has improved.

take bool null indices 1024
                        time:   [1.5549 µs 1.5560 µs 1.5572 µs]
                        change: [-16.412% -15.787% -15.182%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 15 outliers among 100 measurements (15.00%)
  2 (2.00%) low mild
  6 (6.00%) high mild
  7 (7.00%) high severe

take bool null values 1024
                        time:   [1.6879 µs 1.6897 µs 1.6916 µs]
                        change: [-11.554% -11.427% -11.294%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

take bool null values null indices 1024
                        time:   [3.5142 µs 3.5281 µs 3.5463 µs]
                        change: [-16.009% -15.478% -14.934%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  2 (2.00%) low mild
  8 (8.00%) high mild
  2 (2.00%) high severe
```


<!--
Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.
-->

# What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

# Are there any user-facing changes?


<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!---
If there are any breaking changes to public APIs, please add the `breaking change` label.
-->