@Dandandan 2024-10-25
I found this also improves take for primitive / stringview with null values (15 - 20 %, added results in PR comment)